Feat/uiux design

+++ backend/src/main/java/com/panhoramix/backend/service/MediaService.java
@@ -1,6 +1,7 @@
 package com.panhoramix.backend.service;
 
 import com.panhoramix.backend.dto.request.CreateMediaRequest;
+import com.panhoramix.backend.dto.request.UpdateMediaRequest;
 import com.panhoramix.backend.dto.response.MediaPageResponse;
 import com.panhoramix.backend.dto.response.MediaResponse;
 import com.panhoramix.backend.entity.Media;
@@ -170,7 +171,24 @@ public class MediaService {
         Media media = mediaRepository.findById(id)
                 .orElseThrow(() -> new MediaNotFoundException(id));
 
-        fileStorageService.deleteFile(media.getMediaUrl());
+        User currentUser = currentUserService.getCurrentUser();
+
+        if (!media.getUser().getId().equals(currentUser.getId())) {
+            throw new MediaNotFoundException(id);
+        }
+
+        try {
+
+            fileStorageService.deleteFile(media.getMediaUrl());
+
+        } catch (Exception e) {
+
+            throw new RuntimeException(
+                    "Failed to delete media from Cloudflare",
+                    e
+            );
+
+        }
 
         mediaRepository.delete(media);
 
@@ -178,7 +196,7 @@ public class MediaService {
 
     public MediaResponse updateMedia(
             Long id,
-            CreateMediaRequest request) {
+            UpdateMediaRequest request) {
 
         Media media = mediaRepository.findById(id)
                 .orElseThrow(() -> new MediaNotFoundException(id));
@@ -187,7 +205,12 @@ public class MediaService {
         media.setDescription(request.getDescription());
         media.setCategory(request.getCategory());
         media.setVisibility(request.getVisibility());
-        media.setMediaType(request.getMediaType());
+
+        User currentUser = currentUserService.getCurrentUser();
+
+        if (!media.getUser().getId().equals(currentUser.getId())) {
+            throw new MediaNotFoundException(id);
+        }
 
         Media updatedMedia = mediaRepository.save(media);
